앱 분리
