+++ teachercan/students/api.py
@@ -1,0 +1,3 @@
+from ninja import Router
+
+router = Router(tags=["Student"])
