+++ teachercan/users/admin.py
@@ -1,33 +1,7 @@
 from django.contrib import admin
 from django.contrib.auth.admin import UserAdmin
 
-from .models import (
-    User,
-    School,
-    StudentList,
-    Student,
-    Allergy,
-    Column,
-    Row,
-    StudentAllergyRelation,
-)
-
-
-class SchoolInline(admin.TabularInline):
-    model = School
-
-
-class UserInline(admin.TabularInline):
-    model = User
-    extra = 1
-
-
-class SchoolAdmin(admin.ModelAdmin):
-    fieldsets = [
-        ("None", {"fields": ["code", "area_code", "name"]}),
-    ]
-    search_fields = ("name",)
-    inlines = [UserInline]
+from .models import User
 
 
 class BaseUserAdmin(UserAdmin):
@@ -111,20 +85,4 @@ class BaseUserAdmin(UserAdmin):
     readonly_fields = ("id",)
 
 
-class StudentAllergyInline(admin.TabularInline):
-    model = Student.allergy.through
-    extra = 1
-
-
-class StudentAdmin(admin.ModelAdmin):
-    inlines = (StudentAllergyInline,)
-
-
 admin.site.register(User, BaseUserAdmin)
-admin.site.register(School, SchoolAdmin)
-admin.site.register(StudentList)
-admin.site.register(Student, StudentAdmin)
-admin.site.register(Allergy)
-admin.site.register(Column)
-admin.site.register(Row)
-admin.site.register(StudentAllergyRelation)
