+++ teachercan/users/models.py
@@ -10,21 +10,7 @@ from django.contrib.auth.validators import (
     UnicodeUsernameValidator,
 )
 from django.contrib.auth.password_validation import validate_password
-from django.utils import timezone
 from django.utils.translation import gettext_lazy as _
-from django.core.mail import send_mail
-
-
-class School(models.Model):
-    code = models.CharField(max_length=10, null=False, primary_key=True, db_index=True)
-    area_code = models.CharField(max_length=10, null=False)
-    name = models.CharField(max_length=10, null=False)
-
-    class Meta:
-        db_table = "school"
-
-    def __str__(self):
-        return self.name
 
 
 class UserManager(BaseUserManager):
@@ -79,7 +65,7 @@ class User(AbstractBaseUser, PermissionsMixin):
         validators=[nickname_validator],
     )
     school = models.ForeignKey(
-        to="School", null=True, blank=True, on_delete=models.SET_NULL
+        to="schools.School", null=True, blank=True, on_delete=models.SET_NULL
     )
     gender = models.CharField(
         null=True, max_length=2, choices=[("남", "남"), ("여", "여")], default="남"
@@ -101,68 +87,3 @@ class User(AbstractBaseUser, PermissionsMixin):
     @property
     def is_staff(self):
         return self.is_superuser
-
-
-class StudentList(models.Model):
-    name = models.CharField(max_length=15)
-    is_main = models.BooleanField(default=False)
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
-    has_allergy = models.BooleanField(default=False)
-    description = models.CharField(null=True, max_length=200)
-
-    user = models.ForeignKey(User, on_delete=models.CASCADE)
-
-    class Meta:
-        db_table = "student_list"
-
-
-class Allergy(models.Model):
-    code = models.IntegerField(primary_key=True)
-    name = models.CharField(max_length=20)
-
-    class Meta:
-        db_table = "allergy"
-
-
-class StudentAllergyRelation(models.Model):
-    student = models.ForeignKey(to="Student", on_delete=models.CASCADE)
-    allergy = models.ForeignKey(to="Allergy", on_delete=models.CASCADE)
-
-    class Meta:
-        db_table = "student_allergy_set"
-
-
-class Student(models.Model):
-    name = models.CharField(max_length=10)
-    number = models.IntegerField()
-    gender = models.CharField(
-        max_length=2, choices=[("남", "남"), ("여", "여")], default="남"
-    )
-
-    student_list = models.ForeignKey(
-        StudentList, on_delete=models.CASCADE, db_column="list_id"
-    )
-    allergy = models.ManyToManyField(to="Allergy", through="StudentAllergyRelation")
-
-    class Meta:
-        db_table = "student"
-
-
-class Column(models.Model):
-    field = models.CharField(max_length=20)
-    student_list = models.ForeignKey(
-        to="StudentList", on_delete=models.CASCADE, null=True
-    )
-
-    class Meta:
-        db_table = "student_list_column"
-
-
-class Row(models.Model):
-    value = models.CharField(max_length=100)
-    column = models.ForeignKey(to="Column", on_delete=models.CASCADE)
-    student = models.ForeignKey(to="Student", on_delete=models.CASCADE)
-
-    class Meta:
-        db_table = "student_list_row"
